requirements.txt
alembic==1.4.3
aniso8601==8.0.0
beautifulsoup4==4.9.1
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
SQLAlchemy-serializer==1.3.4.2
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
